finist diceEngine test coverage apart from handleRollDice
