+++ src/diceEngine/__tests__/calcDicePool.test.ts
@@ -1,5 +1,4 @@
 import { calcDicePool } from '../diceUtils'
-import { Dice } from '../diceTypes';
 
 describe('calcDicePool', () => {
